Link improvements
- When you hover over an HTML header, we now pop up a link icon holding a link to the
header. Useful to support deep linking to a header.

- When you click on a link that points outside of istio.io, the target is now
opened in a separate tab.

- Improved the coloring of links in the light theme so they're more visible. I was almost not able to
see the difference between normal text and links, now the color delta is greater. Additionally, when
you hover on a link, we now draw an underline.

- Change the color of H1 headers to be black instead of being the same color
as links. They're not links, they shouldn't be that color.